+++ src/com/doubleyellow/scoreboard/prefs/PlayerColorsNewMatch.java
@@ -1,0 +1,7 @@
+package com.doubleyellow.scoreboard.prefs;
+
+public enum PlayerColorsNewMatch {
+    None,
+    TakeFromPreviousMatch,
+    TakeFromPreferences,
+}